Alpha (#2)
* Update duckdns

Added the test stuff for the "run at startup" part.

Can't get it to work though.

* Changed to cron.hourly

Instead of startup script, put script in home directory, and linked in /etc/cron.hourly/

* Create duckdns.sh

Made duckdns.sh so it can be downloaded instead of echoed.

* Update duckdns

Got rid of the clumsy echoing of duckdns.sh script. Changed it to just curl download into the right spot. 

Does the same for the /etc/cron.hourly/duckdns file